# Conflicts:
#	keras_hub/src/models/gemma4/gemma4_causal_lm_preprocessor.py
#	keras_hub/src/models/whisper/whisper_audio_converter.py

+++ keras_hub/src/models/whisper/whisper_audio_converter.py
@@ -233,7 +233,7 @@ class WhisperAudioConverter(AudioConverter):
 
         # Compute the mel spectrogram.
         frames = frame_signal(audio, self.num_fft_bins, self.stride)
-        frames = frames * hann_window(self.num_fft_bins, dtype="float64")
+        frames = frames * hann_window(self.num_fft_bins, dtype=dtype)
         stft = np.fft.rfft(frames, n=self.num_fft_bins, axis=-1)
         magnitudes = np.square(np.abs(stft[:, :-1, :]))
